Error Handling voor abo

+++ lib/mail/sql.php
@@ -28,13 +28,40 @@ function getFormEmails ( $type ) {
 }
 
 function insertIntoMaillist ( $email ) {
+    $insertArray["success"] = FALSE;
+    $insertArray["emailErr"] = "";
+    
     if (!isset($SQL_AVAILABLE)) {
         require $_SERVER['DOCUMENT_ROOT'] . "/lib/account/sql.php";
     }
     $token = randString(10);
     $conn = connectToDatabase();
-    $stmt = $conn->prepare ( "INSERT INTO mail_list ( email, token ) VALUES ( ?, ? )" );
-    $stmt->execute ( array ( $email, $token ) );
+    
+    $stmt = $conn->prepare ( "SELECT email FROM mail_list WHERE email = ?" );
+    $stmt->execute( array ( $email ) );
+    $row = $stmt->fetch();
+    if ( empty($email) ) {
+        $insertArray["emailErr"] = "Voer A.U.B. een email adres in!";
+        return $insertArray;
+    }
+    if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
+        $insertArray["emailErr"] = "Email is ongeldig.";
+        return $insertArray;
+    }
+    if ( $row[0] === $email ){
+        $insertArray["emailErr"] = "Dit email adres bestaat al!";
+        return $insertArray;
+    }
+    else {
+        $stmt = $conn->prepare ( "INSERT INTO mail_list ( email, token ) VALUES ( ?, ? )" );
+        if ( $stmt->execute ( array ( $email, $token ) ) ) {
+            $insertArray["success"] = TRUE;
+            return $insertArray;
+        }
+        else {
+            return $insertArray;
+        }
+    }
 }
 
 function getAllMallist ( ) {
@@ -61,9 +88,17 @@ function unsubscribe ( $token, $email ) {
     }
     $conn = connectToDatabase();
     $checkToken = getMaillistToken($conn, $email);
+    if ( empty($checkToken) ) {
+        return FALSE;
+    }
     if ( $token === $checkToken[0][0] ) {
         $stmt = $conn->prepare( "DELETE FROM mail_list WHERE token = ? AND email = ?");
-        $stmt->execute ( array ( $token, $email ) );
+        if ( $stmt->execute ( array ( $token, $email ) ) ) {
+            return TRUE;
+        }
+        else {
+            return FALSE;
+        }
     }
     else {
